requirements.txt
Flask
numpy
snowland-smx
mysql-connector-python
Flask-SocketIO
pycryptodomex